requirements.txt
numpy~=1.24.3
pathlib~=1.0.1
yaml~=0.2.5
pyyaml~=6.0
matplotlib~=3.7.2
pandas~=2.0.3
seaborn~=0.12.2
pillow~=10.2.0
scipy~=1.11.1
tqdm~=4.65.0
requests~=2.31.0
sphinx~=5.0.2
tabulate~=0.8.10
psutil~=5.9.0
setuptools~=68.0.0
datasets~=2.12.0
packaging~=23.1
pycocotools~=2.0.7
cloudpickle~=2.2.1